requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.33.0
azure-identity==1.21.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.9.0
msal==1.32.0
msal-extensions==1.3.1
openai==1.71.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
pillow==11.2.1
pip==25.0.1
pypdfium2==4.30.1
setuptools==65.5.0